backend/requirements.txt
aiobotocore==2.15.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cors==0.7.0
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
boto3==1.35.14
botocore==1.35.16
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
chartjs==1.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
duckdb==1.1.3
eventlet==0.36.1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gevent==24.2.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_resources==6.4.4
ipython==8.18.1
jedi==0.19.1
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
multidict==6.0.5
numpy==2.0.2
openai==1.42.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pprintpp==0.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.2.1
python-socketio==5.4.1
pytz==2024.1
RapidFuzz==3.10.0
requests==2.32.3
rsa==4.9
s3fs==2024.9.0
s3transfer==0.10.2
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
uvloop==0.20.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.3
